UV/simulation/pyproject.toml
[project]
name = "simulation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.17",
    "tavily-python>=0.7.6",
]
